+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 - Add Pipeline `select` method
 - Add Pipeline `filter` & `filter_col` methods
 - Add Pipeline `from_pipelines` constructor for merging pipelines together
+- Add `count` transformer
 - Remember row order in transform_into
 - Include source index in errors